The caller hears a burst of noise when a call waiting call is answered and connects.
1) The dialing, digital, and outgoing flags need to be transferred when 
ISDN B channels are switched.  (The noise burst was not heard on trunk 
level code because the dialing flag was not transferred when B channels 
were switched.) 

2) The dialing flag is used as a receive squelch for ISDN calls until 
connected.  Do not clear the dialing flag until the call is connected or a 
progress indication ie indicates inband information is available.  (The 
noise burst was heard on C.3 level code because the dialing flag was 
cleared when the PROCEEDING message came in.)