Fix a race condition found during testing Call completion.
If a single monitored device had two requests made to it, then it was
possible for a device state change to be "missed" after the first recall
has completed. The reason for this is that the generic_monitor_instance
for the first recall was still in the list of instances for the particular
device. The result was that when the device became available, we would signal
its availability to the incorrect core_id. This change fixes the problem
by adding an integer which is set immediately false once we signal an
availability change on the device.