Changes to acl.c to deal with both IPv4 and IPv6 host access rules.
Since when applying an HA, it is not clear if the address has been
given by the kernel or parsed from something like a SIP packet, care
needs to be taken to be sure rules are applied accurately. The way I've
done this is to internally store all host access rules as IPv6 addresses.
If a rule is written to apply to an IPv4 address, we store it as an
IPv4-mapped address.

The code does not compile at the moment. While there are likely some errors
in acl.c, the bigger problem is that there is code which has not been
changed to use the new ast_append_ha and ast_apply_ha calls.