Make modifications to my aclv6-branch based on discussion on the -dev list.
The main change made here was to not store all ACLs internally as IPv6 addresses.
Instead, there is a flag in the ast_ha struct that indicates what the address scheme
used in the particular rule. This way, IPv4 ACLs only apply to IPv4 addresses and
IPv6 ACLs only apply to IPv6 addresses.

IPv4-mapped addresses in all cases are stored and treated as IPv4 addresses. In
order to facilitate this, I had to make the ast_sockadr_ipv4_mapped() function
in netsock2.c public. I added an appopriate warning to the doxygen to indicate
that it's not likely to be needed very often like several of the other functions.