Add a simple AMI client web page
This patch uses the XML docs to parse all of the available AMI commands
and allows you to enter the command name and be presented with a form with
the available fields. You can then rapidly tab through the fields and submit
the command and view the response. It is much faster/easier than having to
use telnet for testing purposes.