Create branch for solving issue 18336.
The idea here is to allow for the CCSS core to notify
a CC agent that a failure has occurred when CC is requested.

The plan is to modify the "ack" callback to become a "respond"
callback instead that has a parameter indicating either a
successful response or a specific type of failure that may
need to be communicated to the requester.

The reason for creating a branch for fixing this is that fixing
the bug will be a collaborative effort since changes will be needed
in the core as well as all CC agent implementations.