Fix voicemail sequencing for file based storage.
A previous change was made to account for when the number of voicemail messages
exceeds the max limit to be handled properly, but it caused gaps in the messages
to not be properly handled. This has now been resolved.

In later non 1.4 branches, it appears that resequencing wasn't even occurring
due from what appears and accidental code removal.

(closes issue #18498)
Reported by: JJCinAZ
Patches: 
      bug18498v2.patch uploaded by jpeeler (license 325)

(closes issue #18486)
Reported by: bluefox
Patches: 
      bug18486.patch uploaded by jpeeler (license 325)